Move FastAPI client-related functions to service-specific utilities
